requirements.txt
google-genai>=1.31.0
pdf2image>=1.17.0
pillow>=11.3.0
pandas>=2.3.0
openpyxl>=3.1.5
pydantic>=2.11.0
tqdm>=4.67.0
loguru>=0.7.3
python-dotenv>=1.1.1
tenacity>=9.1.2
sympy>=1.14.0
rapidfuzz>=3.14.1
pytest>=8.4.1
# Web frontend
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
alembic>=1.14
python-multipart>=0.0.20
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
apscheduler>=3.11
jinja2>=3.1
markdown>=3.7
